last updated

+++ src/main/java/com/zidani/gestioncv/experienceManagment/ExperienceController.java
@@ -1,12 +1,16 @@
 package com.zidani.gestioncv.experienceManagment;
 
+import com.zidani.gestioncv.authenticationManagment.tokenManagement.AuthorizationService;
+import com.zidani.gestioncv.config.JwtService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 @Tag(name = "Experience", description = "Experience Management Api")
@@ -16,6 +20,8 @@ import org.springframework.web.bind.annotation.*;
 public class ExperienceController {
 
     private final ExperienceService experienceService;
+    private final JwtService jwtService;
+    private final AuthorizationService authorizationService;
 
     @PostMapping("/{cvId}")
     @Operation(summary = "Add a new experience to the CV", description = "Endpoint to add a new experience to the CV")
@@ -24,12 +30,20 @@ public class ExperienceController {
             @ApiResponse(responseCode = "404", description = "CV not found"),
             @ApiResponse(responseCode = "500", description = "Internal server error")
     })
-    public ResponseEntity<Long> addExperience(
+    public ResponseEntity<?> addExperience(
             @PathVariable Long cvId,
-            @RequestBody ExperienceRequest experienceRequest
+            @RequestBody ExperienceRequest experienceRequest,
+            HttpServletRequest servletRequest
     ) {
-        Long experienceId = experienceService.addExperience(cvId, experienceRequest);
-        return ResponseEntity.ok(experienceId);
+        String username = jwtService.extractUsernameAndBearerToken(servletRequest);
+
+        if (authorizationService.hasExperienceRight(username, cvId)) {
+            Long experienceId = experienceService.addExperience(cvId, experienceRequest);
+            return ResponseEntity.ok(experienceId);
+        }
+        return ResponseEntity
+                .status(HttpStatus.FORBIDDEN)
+                .body("Forbidden ");
     }
 
     @PutMapping("/{id}")
@@ -39,13 +53,20 @@ public class ExperienceController {
             @ApiResponse(responseCode = "404", description = "Experience not found"),
             @ApiResponse(responseCode = "500", description = "Internal server error")
     })
-    public ResponseEntity<Experience> updateExperience(
+    public ResponseEntity<?> updateExperience(
             @PathVariable Long id,
-            @RequestBody ExperienceRequest experienceRequest
+            @RequestBody ExperienceRequest experienceRequest,
+            HttpServletRequest servletRequest
     ) {
-        Experience updatedExperience = experienceService.updateExperience(id, experienceRequest);
-        return ResponseEntity.ok(updatedExperience);
-    }
+        String username = jwtService.extractUsernameAndBearerToken(servletRequest);
+
+        if (authorizationService.hasExperienceRight(username, id)) {
+            Experience updatedExperience = experienceService.updateExperience(id, experienceRequest);
+            return ResponseEntity.ok(updatedExperience);
+        }
+        return ResponseEntity
+                .status(HttpStatus.FORBIDDEN)
+                .body("Forbidden ");    }
 
     @DeleteMapping("/{id}")
     @Operation(summary = "Delete an existing experience", description = "Endpoint to delete an existing experience")
@@ -54,12 +75,19 @@ public class ExperienceController {
             @ApiResponse(responseCode = "404", description = "Experience not found"),
             @ApiResponse(responseCode = "500", description = "Internal server error")
     })
-    public ResponseEntity<Void> deleteExperience(
-            @PathVariable Long id
+    public ResponseEntity<?> deleteExperience(
+            @PathVariable Long id,
+            HttpServletRequest servletRequest
     ) {
-        experienceService.deleteExperience(id);
-        return ResponseEntity.noContent().build();
-    }
+        String username = jwtService.extractUsernameAndBearerToken(servletRequest);
+
+        if (authorizationService.hasExperienceRight(username, id)) {
+           experienceService.deleteExperience(id);
+            return ResponseEntity.ok("Experience supprimé");
+        }
+        return ResponseEntity
+                .status(HttpStatus.FORBIDDEN)
+                .body("Forbidden ");    }
 
     @GetMapping("/searchByExperienceTitle")
     public Page<?> searchByExperienceTitle(
